api version 2 with jwt

+++ api/Controllers/StockController.cs
@@ -4,9 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using api.Data;
 using api.Dtos.Stock;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using api.Repository;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -25,17 +27,24 @@ namespace api.Controllers
             _stockRepository = stockRepository;
         }
 
+
         [HttpGet("all")]
-        public async Task<IActionResult>  GetAll(){
-            var stocks = await _stockRepository.GetAllAsync();
+        [Authorize]
+        public async Task<IActionResult>  GetAll([FromQuery]QueryObject queryObject){
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            var stocks = await _stockRepository.GetAllAsync(queryObject);
              var stockDto =    stocks.Select(s => s.ToSTockDto());
             return Ok(stocks);
 
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
+        [Authorize]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await  _stockRepository.GetByIdAsync(id);
             if(stock == null){
                 return NotFound();
@@ -45,8 +54,11 @@ namespace api.Controllers
         }
 
         [HttpPost("createStock")]
+        [Authorize]
         public async Task<IActionResult> Create([FromBody] CreateStockDto stockDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockmodel = stockDto.toStockFromCreateDto();
             await _stockRepository.CreateAsync(stockmodel);
             return CreatedAtAction(nameof(GetById), new {id = stockmodel.Id}, stockmodel.ToSTockDto());
@@ -54,9 +66,12 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("update/{id}")]
+        [Route("update/{id:int}")]
+        [Authorize]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateStockRequestDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var Stock = await _stockRepository.UpdateAsync(id,updateStockRequestDto);
             if(Stock == null){
                 return NotFound();
@@ -66,9 +81,12 @@ namespace api.Controllers
             
         }
 
-        [HttpDelete("delete/{id}")]
+        [HttpDelete("delete/{id:int}")]
+        [Authorize]
         public async Task<IActionResult> Delete([FromRoute]int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
             var Stock = await _stockRepository.RemoveAsync(id);
             if(Stock == null){
                 return NotFound();
